Major refactorint Advanced Query
