+++ src/android/com/outsystems/plugins/healthfitness/store/AdvancedQueryParameters.kt
@@ -7,7 +7,8 @@ data class AdvancedQueryParameters (
     @SerializedName("Variable") val variable : String,
     @SerializedName("StartDate") val startDate : Date,
     @SerializedName("EndDate") val endDate : Date,
-    @SerializedName("TimeUnit") val timeUnit : String?,
+    @SerializedName("TimeUnit") val timeUnit : String? = null,
     @SerializedName("TimeUnitLength") var timeUnitLength : Int? = 1,
-    @SerializedName("OperationType") val operationType : String? = EnumOperationType.RAW.value
+    @SerializedName("OperationType") val operationType : String? = EnumOperationType.RAW.value,
+    @SerializedName("Limit") val limit : Int? = null
 )
